Using FOR1D directly. Still wrong results!
